+++ weblate/formats/base.py
@@ -31,6 +31,7 @@ from sentry_sdk import add_breadcrumb
 from weblate_language_data.countries import DEFAULT_LANGS
 
 from weblate.utils.hash import calculate_hash
+from weblate.utils.state import STATE_TRANSLATED
 
 EXPAND_LANGS = {code[:2]: f"{code[:2]}_{code[3:].upper()}" for code in DEFAULT_LANGS}
 
@@ -143,12 +144,8 @@ class TranslationUnit:
         """Set translation unit target."""
         raise NotImplementedError()
 
-    def mark_fuzzy(self, fuzzy):
-        """Set fuzzy flag on translated unit."""
-        raise NotImplementedError()
-
-    def mark_approved(self, value):
-        """Set approved flag on translated unit."""
+    def set_state(self, state):
+        """Set fuzzy /approved flag on translated unit."""
         raise NotImplementedError()
 
 
@@ -465,7 +462,7 @@ class TranslationFormat:
             # Unmark unit as fuzzy (to allow merge)
             set_fuzzy = False
             if fuzzy and unit.is_fuzzy():
-                unit.mark_fuzzy(False)
+                unit.set_state(STATE_TRANSLATED)
                 if fuzzy != "approve":
                     set_fuzzy = True