+++ docs/changes.rst
@@ -19,6 +19,7 @@ Not yet released.
 * Password hashing now uses Argon2 by default.
 * Simplified progress bars indicating translation status.
 * Renamed :ref:`addon-weblate.consistency.languages` to clarify the purpose.
+* Fixed saving string state to XLIFF.
 
 `All changes in detail <https://github.com/WeblateOrg/weblate/milestone/61?closed=1>`__.